Modify playRound function to call the play functions by itself
The playRound function used to receive the plays from both the user
and the computer, and played the round based on those parameters.
Now, it calls the plays by itself, so it doesn't need to receive any
parameters.

This change is done so the "tie" result can call itself in a future
change.