Modify the tie result of the function playRound
If the playRound function resulted in a tie, it would treat it the same as
the other results and return a string.

Since the game of Rock Paper Scissors is normally used to resolve disputes,
a tie is not an acceptable result.
So, the change was made that if a tie occurs, the function calls itself
until either a Win or a Loss is returned.