Create helper functionn checkResult
This function checks for the result of the plays.
Receives two parameters, the player's play and the computer's play,
both in strings.
Then, converts the string to a numerical value so it can verify who won
without working through every possible variation of the game.

This function is called inside the playRound function.